requirements.txt
fastapi
uvicorn
openai>=1.0.0
python-docx
pandas
matplotlib
python-multipart
openpyxl
